Add files via upload
The "cards.zip" file in the assets of the game is a compressed archive containing the digital representations of playing cards, from Ace (1) to King (13). These card images are essential components of the game, visually representing the cards that players will encounter and predict during gameplay. The zip file ensures efficient storage and easy access to the card assets, enhancing the game's performance and visual appea