Add files via upload
The "MainStoryboard.storyboard" file serves as the central blueprint for the user interface and navigation flow of the Higher-Lower game. It visually represents the arrangement of screens, buttons, and views that players interact with as they progress through the game. This essential component of the game's development defines the structure and layout, ensuring a seamless and intuitive user experience.