Add files via upload
The "LaunchScreen.storyboard" is the opening act of the Higher-Lower game, setting the stage for the player's adventure. This storyboard presents a visually appealing and engaging welcome screen that appears when the game is launched. It introduces the game's branding, logo, or initial message, creating a captivating first impression that draws players into the gaming experience from the moment they start the app.