Add files via upload
The "HighLow.xcodeproj.zip" file is an archived Xcode project that encapsulates the entire development environment and settings for the Higher-Lower game. This compressed archive includes the project configuration, source code, assets, and dependencies required to build and run the game within the Xcode development environment. It offers a convenient way to package and share the complete project with collaborators or to store a snapshot of the project at a particular stage of development, ensuring that the game can be easily worked on and reproduced